Incorporated GUI
Created a three page GUI that tracks button clicks and the time that the button is pressed and records it to a .csv file. I currently only have it working with one button recording time, and my next step is to make it calculate the elapsed time and then have it return the input it received.